change data types

+++ web/src/main/kotlin/gruzilkin/iot/queue/SensorDataEvent.kt
@@ -1,8 +1,10 @@
 package gruzilkin.iot.queue
 
+import java.math.BigDecimal
+
 data class SensorDataEvent(
     val deviceId: Long,
     val sensorName: String,
-    val sensorValue: Double,
+    val sensorValue: BigDecimal,
     val receivedAt: Long
 )
